WIP Implement Reset ALL function.
I try to chew my teeth into that code. This is a work in progres as i work on my lsl skills. 

The idea for this comit is to implement a ResetALL function to reset the Collar to its default settings. The "reset ALL" button ore better the "Runaway" Button dos not automaticaly reset the Collar. This is expectet behavior so the sub can flee. But the Resetfunction is much needet.