Medea Destiny: major rlvex menu overhaul
Medea: "I've overhauled the menu appearance as it wasn't really that
easy to understand, following the RLV terminology which is a bit of a
headache for scripters, let alone users. Descriptions are clearer and
now represent the current state, and we use unicode ☐ and ☒ to indicate
selected/deselected state on the buttons, common with unicode
enhancements to other menus. ALL buttons indicators swapped to make
clearer the effect of pressing them, and the person to whom the
exceptions will be added/removed is now named at the top of the menu.
Finally, I've cleared up remenuing a little. When someone who is not an
owner/secowner is set to default and removed from the list, we now
return the personlist, and when a new person is added, we remenu."