Medea Destiny in-world; texture
Medea: "Added support for texture notecard named textures_custom which
will be read instead of the standard texture notecard if present. Also
put MENUNAME_RESPONSE into onrez in place of reset. This menu building
system is... uh... yeah."

Karo: "removed llToLower() in GetLongName(), GetElementHasTexs() and
BuildTexButtons() functions. This commit makes textureable elements case
sensitive; we need that if textures are exclusive to specific elements
on the collar otherwhise those textures would have to be named like the
element but in lower caps"

Medea: "Flip side, now it relies on them being the same. Risk of
breaking some existing content here? I'm inclined to say it would be
better to put llToLower on both sides of the test comparisons, no?"