Medea Destiny in-world; rlvex (fixes #45)
Medea: Bugfixes -  1. Reported issue, message about UUID not found when
trying to add other. I haven't found this to be repeatable, but I have
managed to get it to do it once. Increased llRequestAgentInfo timer to 4
seconds from 0.5, because 0.5 may simply be too short to reliably get a
response from the dataserver, particularly in a laggy sim. Hopefully
this will remove the issue, but let's monitor this.  2. Reported issue
in secowner exceptions not being set. Typo in settings was looking for
auth_secowner rather than auth_secowners in LM_SETTINGS_RESPONSE, so
secowners list was not getting filled.