fixes #64
I'm basically just reverting one change, added the reset back and
doubled up the request for permissions. That fixes our problem with
losing perms to animate. It's a bizarre way to go about it and the reset
feels out of place but I have yet to see an alternative fix that works?