re-added missing attributions and changelog
The script was missing attributions which must have been lost while work
was done on it in-world; this script was originally created by WhiteFire
Sondergaard based on the plugin template at the time by Cleo Collins.
Kadah Coba, Satomi Ahn and various others contributed improvements and
patches to this script.