prepping 3.955; Bundles I
First here is addressing the realleash plugins that have to be
uninstalled or we would double up the same the normal leash does now.

Second is the updater that caused more confusion if handed from
contents, so for now, let's have Get Updater inform users /where/ to get
the actual update rather than receive a package manager style updater
from contents. I will be working on the necessary changes in OpenCollar
- update to have it shipped with 3.955.

Third is the keys/keyholder/keyholder plugin. Putting it back in
received no feedback, the issues on it received no response and the
problem with missing attributions has been addressed properly. I believe
at this point it will be best to stick to our popular set of official
plugins and the two "missing" rlv features. I will contact Whitefire
Sondergaard and ask for a link and slurl where interested people could
acquire the plugin from the original author. These links can be blogged
with the announcement for beta 3.955!