fix #277 by littlemousy
I have received this as in-world contribution by littlemousy. She also
mentioned that this is a really good example of saving code from the
"textkey" change. I feel the same because I have been able to understand
it this way too now and for those interested, this could be reference to
gain better understanding about it all.