Create OpenCollar - styles.lsl
want to quickly change styles collar? Now it's easy!

in Appearance menu *Styles* , with "Dump Style" button dump current appearance that set using "texture", "color" ("shine"), and write it as "your style" in 'styles' notecard in the collar.
And this style will appear in the Styles menu for further use.