resurrect view section
Alright, so we said we make a test to see if the minimum required
version of RLV is used and depending on that have several options how to
handle things if the version cannot play camera/view features:

a) We use an alternative menu informing the minimum required RLV version
is not used.

b) We don't show the button in the first place.

c) We show the button but instead of letting it give the menu we do a
notify to inform the minimum required RLV version is not used.