standardize DEPRECATED bundle
This installer bundle is called "Cruft" and tagged DEPRECATED. When we
perform an installation, bundles tagged DEPRECATED will always be
removed and cannot be unticked in selection menus. A good choice for
this type of bundle could be everything that makes the device sluggish,
bloated or inoperational.

Example:
ITEM|pr0n
SCRIPT|mcafee