standardize textures card
In this card we can specify textures by category and name. If a texture
follows the format "Name,UUID" then this texture will be made available
to all elements of the device that are allowed to be textured. If the
format is "Category~Name,UUID" the texture will only be made available
to elements described as such category.